Add test csv builder verify
