--- test/unit/ggrc/bulk_operations/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (C) 2019 Google Inc.
-# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>]
